Move mockServer into a separate file.
